% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdhg.R
\name{adaptive_pdhg}
\alias{adaptive_pdhg}
\title{Adaptive PDHG algorithm for group GMC}
\usage{
adaptive_pdhg(
  beta,
  v,
  y,
  X,
  Z,
  group,
  K,
  lambda,
  maxiters = 500,
  tol = 0.001,
  adaptive = TRUE,
  backtrack = TRUE,
  L,
  tau,
  sigma,
  a = 0.5,
  eta = 0.95,
  Delta = 2,
  gamma = 0.75,
  b = 0.95,
  stopRule = "ratioResidual",
  ShowTime = FALSE
)
}
\arguments{
\item{beta}{The initial value for variable beta. If missing, set it as a zero vector.}

\item{v}{The initial value for variable v. If missing, set it as a zero vector.}

\item{y}{The response vector.}

\item{X}{The design matrix in group GMC.}

\item{Z}{The Z matrix in group GMC (Z = lambda*B'B/n).}

\item{group}{A vector with the same length as beta. Each component indicates which group
he corresponding covariate variable belongs to.}

\item{K}{The weight vector for the group penalty or called group multiplier.}

\item{lambda}{The regularization parameter in group GMC.}

\item{maxiters}{The max number of iterations.}

\item{tol}{A tolerance for early stopping.}

\item{adaptive}{If 'TRUE', use adaptive method.
If 'FALSE', turn off the adaptivity and users must set tau and
 sigma as the setpsize for the classical PDHG.}

\item{backtrack}{If 'TRUE', use backtracking method.}

\item{L}{The reciprical spectral radius of Z'Z. If missing, the algorithm will give an approcimate value.}

\item{tau}{The intial stepsize for the primal variables.
If missing, the algorithm will initialize it as the squared root of L.}

\item{sigma}{The intial stepsize for the dual variables.
If missing, the algorithm will initialize it as L/tau.}

\item{a}{Intial adaptive update strength for stepsizes.}

\item{eta}{How fast does the adaptivity level decay.}

\item{Delta}{Update stepsizes when primal/dual ratio exceeds Delta.}

\item{gamma}{Used to determine when need to backtrack to maintain positivity conditions.}

\item{b}{Adaptivity parameter used for backtracking update.}

\item{stopRule}{Three choices in total. Default is ratioResidual.
"ratioResidual": primal/maxPrimal<tol & dual/maxDual < tol.
"residual": primal < tol & dual < tol.
"iteration" iter > maxiters.}

\item{ShowTime}{Whether to show the time cost for computing the solution.}
}
\description{
Adaptive PDHG algorithm for group GMC
}
