% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.GMC.R
\name{cv.GMC}
\alias{cv.GMC}
\title{Cross validation for (group) GMC}
\usage{
cv.GMC(
  y,
  X,
  alpha = 0.8,
  group = 1:ncol(X),
  group.multiplier,
  nlambda = 50,
  lambdaSeq,
  lambda.min = {     if (nrow(X) > ncol(X))          0.001     else 0.05 },
  log.lambda = TRUE,
  nfolds = 5,
  algorithm = "PDHG",
  maxiters = 100,
  tol = 1e-06,
  seed = 1234,
  trace = FALSE
)
}
\arguments{
\item{y}{The response vector.}

\item{X}{The design matrix.}

\item{alpha}{The convexity-preserving parameter taking on a value between 0 and 1. Default is 0.8.}

\item{group}{A vector of length ncol(X). Each component indicates which group the corresponding covariate variable belongs to.}

\item{group.multiplier}{The weight vector for the group GMC penalty.}

\item{nlambda}{The number of values for the regularization parameter. Default is 50. The function will automatically computes
a grid of 'nlambda' lambda values with the first one producing a zero solution.}

\item{lambdaSeq}{A user-specified sequence of decrasing lambda values. It is typically left unspecified and let the function automatically
computes a grid of 'nlambda' lambda values.}

\item{lambda.min}{The smallest lambda value, as a fraction of the maximum lambda.
Default is .001 if the number of observations is larger than the number of covariates and .05 otherwise.}

\item{log.lambda}{Whether to compute the grid values of lambda on log scale (default) or linear scale. Default is TRUE.}

\item{nfolds}{The number of cross validation folds. Default is 5.}

\item{algorithm}{Which algorithm to use for solving the group GMC problem. PDHG (default) or FB.
We recommend using PDHG for large value and FB for small alpha.}

\item{maxiters}{The maximal number of iterations. Typically, PDHG only needs 100 but FB needs 1e4.}

\item{tol}{A tolerance for early stopping. Default is 1e-6.}

\item{seed}{Random seed for reproduccible results.}

\item{trace}{Whether to track the progress of the cross validation. Default is FALSE.}
}
\value{
\code{cve} The cross-validation error for each value of lambda, averaged over nfolds.

\code{cvse} The standard error associated with each value of cve.

\code{lambdaSeq} The sequence of lambda values along which the cross-validation error was calculated.

\code{lambda_min} The value of lambda with the minimum cross-validation error.

\code{min} The index of lambda corresponding to lambda_min.

\code{lambda_1se} The value of lambda selected by the 1se rule.

\code{min_1se} The index of lambda corresponding to lambda_1se.
}
\description{
\code{cv.GMC} conduct cross validation for (group) GMC penalized regression over a grid of values for the regularization parameter lambda.
}
\author{
Xiaoqian Liu
}
