% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.GMC.R
\name{cv_fold}
\alias{cv_fold}
\title{Cross-validation fold}
\usage{
cv_fold(i, y, X, fold, cv.args)
}
\arguments{
\item{i}{The i-th fold}

\item{y}{The response vector}

\item{X}{The design matrix}

\item{fold}{The fold indices}

\item{cv.args}{The arguments}
}
\description{
\code{cv_fold} fit the model for the i-th fold
}
