% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{mmprod_C}
\alias{mmprod_C}
\title{The matrix-matrix multiplication using C code}
\usage{
mmprod_C(A, B)
}
\arguments{
\item{A}{one matrix multiplier}

\item{B}{tne other matrix multiplier}
}
\description{
\code{mmprod_C} computes the computes the product of two matrices C=AB
}
