% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{mtvprod_C}
\alias{mtvprod_C}
\title{The matrix-vector multiplication using C code}
\usage{
mtvprod_C(A, x)
}
\arguments{
\item{A}{the matrix multiplier}

\item{x}{the vector multiplier}
}
\description{
\code{mtvprod_C} computes the computes the product of b=A'x
}
