% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{mvprod_C}
\alias{mvprod_C}
\title{The matrix-vector multiplication using C code}
\usage{
mvprod_C(A, x)
}
\arguments{
\item{A}{the matrix multiplier}

\item{x}{the vector multiplier}
}
\description{
\code{mvprod_C} computes the computes the product of b=Ax
}
