% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{proxL2}
\alias{proxL2}
\title{The proximal operator of the L2 norm}
\usage{
proxL2(x, tau)
}
\arguments{
\item{x}{the vector}

\item{tau}{the parameter}
}
\description{
\code{proxL2} computes the proximal operator of the L2 norm, tau*\| x \|_2
}
